Add a custom analyzer with libtooling library
The goal of the implementation was to implement a syntax analyzer that can recognize any C++ function declaration without semantic analysis. The C++ standard means that it is possible. But with clang this is not possible, because syntax and semantic analysis are strongly connected.

The result of this commit is a custom semantic analyzer based on the libtooling library. Libtooling is designed to use real files. I implement a virtual file to use the source code from the input (stdin).

The output is really simple. If defined, the name of a function declaration is printed or the declaration of a class, structure or union is dumped.